v1.0.2 Release Support fields without any value. Add constructor, getter and setter for javaclass. Add more methods on JavaFields Change multilinecomentary from immutable to mutable array Fix static field names

+++ src/test/java/ink/reactor/fission/java/JavaInterfaceTest.java
@@ -23,7 +23,7 @@ public class JavaInterfaceTest {
             import lombok.NonNull;
 
             public interface TestClass {
-                int TEST_CONSTANT = 1;
+                short TEST_CONSTANT = 1;
 
                 @Deprecated(since = "2025", forRemoval = true)
                 default boolean isPublic(final @NonNull JavaVisibility visibility) throws Exception {
